Move convert_to_qimage from core to new src/ui/qt_utils
Create new src/ui package with qt_utils module containing convert_to_qimage
function. Remove convert_to_qimage and QImage import from core/image_processing.py
to ensure core module remains Qt-free. Update handlers/display.py to import
convert_to_qimage from ui.qt_utils. Core module now only exports pure-Python
image processing functions.
